requirements.txt
SQLAlchemy~=2.0.28
alembic~=1.13.1
fastapi~=0.110.0
pydantic~=2.6.4
pydantic-settings~=2.2.1
asyncpg==0.29.0
uvicorn==0.29.0
pandas~=2.2.1
scikit-surprise~=1.1.4